Refactor: remove unnecessary unboxing in Ints.tryParse

+++ guava/src/com/google/common/primitives/Ints.java
@@ -877,7 +877,7 @@ public static OptionalInt indexOfOptional(int[] array, int target) {
    */
   public static @Nullable Integer tryParse(String string, int radix) {
     Long result = Longs.tryParse(string, radix);
-    if (result == null || result.longValue() != result.intValue()) {
+    if (result == null || result != result.intValue()) {
       return null;
     } else {
       return result.intValue();
